[Refactor] move Coordinator to separate fine
