+++ iOS-FakeNFT-Extended/UI/User/UserAgreementView.swift
@@ -37,7 +37,7 @@ struct UserAgreementView: View {
     }
 }
 
-private struct WebView: UIViewRepresentable {
+struct WebView: UIViewRepresentable {
     let url: URL
     @Binding var isLoading: Bool
     
@@ -54,22 +54,6 @@ private struct WebView: UIViewRepresentable {
     
     func makeCoordinator() -> Coordinator {
         Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, WKNavigationDelegate {
-        var parent: WebView
-        
-        init(_ parent: WebView) {
-            self.parent = parent
-        }
-        
-        func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-            parent.isLoading = false
-        }
-        
-        func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: any Error) {
-            parent.isLoading = false
-        }
     }
 }
